requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
bottle==0.12.25
Django
django-bootstrap-v5==1.0.11
django-ckeditor==6.7.0
django-environ==0.11.2
django-jazzmin==2.6.0
django-js-asset==2.2.0
fontawesomefree==6.5.1
Pillow==10.1.0
pip-autoremove==0.10.0
psycopg2-binary==2.9.9
setuptools==69.0.2
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.9.0
whitenoise==6.6.0
